refactor: Rename `localStorage` package to `local`
Moves all files from the `core.data.localStorage` package to `core.data.local` and updates all corresponding import statements.
