+++ app/src/androidTest/java/com/github/warnastrophy/core/ui/features/health/HealthCardRepositoryProviderTest.kt
@@ -3,7 +3,7 @@ package com.github.warnastrophy.core.ui.features.health
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.warnastrophy.core.data.localStorage.LocalHealthCardRepository
+import com.github.warnastrophy.core.data.local.LocalHealthCardRepository
 import com.github.warnastrophy.core.data.provider.HealthCardRepositoryProvider
 import com.github.warnastrophy.core.data.repository.HybridHealthCardRepository
 import com.google.firebase.auth.FirebaseAuth
