+++ app/src/main/java/com/github/warnastrophy/core/data/local/StorageResults.kt
@@ -1,4 +1,4 @@
-package com.github.warnastrophy.core.data.localStorage
+package com.github.warnastrophy.core.data.local
 
 import com.github.warnastrophy.core.model.HealthCard
